完成对TasksPage的改造
