+++ SimpleFFmpegGUI.WPF/Pages/TasksPage.xaml.cs
@@ -2,7 +2,6 @@
 using FFMpegCore;
 using FFMpegCore.Enums;
 using FFMpegCore.Pipes;
-using FzLib;
 using FzLib.WPF;
 using Mapster;
 using Microsoft.Extensions.DependencyInjection;
@@ -10,15 +9,13 @@ using SimpleFFmpegGUI.Dto;
 using SimpleFFmpegGUI.Manager;
 using SimpleFFmpegGUI.Model;
 using SimpleFFmpegGUI.WPF.Model;
+using SimpleFFmpegGUI.WPF.ViewModels;
 using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.ComponentModel;
 using System.Diagnostics;
 using System.Drawing;
 using System.Globalization;
 using System.IO;
-using System.Linq;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
@@ -37,46 +34,6 @@ using Size = System.Drawing.Size;
 
 namespace SimpleFFmpegGUI.WPF.Pages
 {
-    public class TasksPageViewModel : INotifyPropertyChanged
-    {
-        public TasksPageViewModel()
-        {
-            AllTasks.PropertyChanged += AllTasks_PropertyChanged;
-            RefreshPages();
-        }
-
-        private void AllTasks_PropertyChanged(object sender, PropertyChangedEventArgs e)
-        {
-            RefreshPages();
-        }
-
-        private void RefreshPages()
-        {
-            var pages = Enumerable.Range(0, AllTasks.PageCount)
-                 .Select(p => new
-                 {
-                     Label = "第" + (p + 1) + "页",
-                     Value = p,
-                     Value1 = p + 1
-                 });
-            if (Pages == null || pages.Count() != Pages.Cast<object>().Count())
-            {
-                Pages = pages;
-            }
-        }
-
-        private IEnumerable pages;
-
-        public IEnumerable Pages
-        {
-            get => pages;
-            set => this.SetValueAndNotify(ref pages, value, nameof(Pages));
-        }
-
-        public event PropertyChangedEventHandler PropertyChanged;
-
-        public AllTasks AllTasks => App.ServiceProvider.GetService<AllTasks>();
-    }
 
     /// <summary>
     /// Interaction logic for TasksPage.xaml
@@ -85,19 +42,10 @@ namespace SimpleFFmpegGUI.WPF.Pages
     {
         public TasksPageViewModel ViewModel { get; set; }
 
-        //private TimeSpan length;
-        //private double fps;
-
-        public TasksPage(TasksPageViewModel viewModel)
+        public TasksPage()
         {
             InitializeComponent();
-            ViewModel = viewModel;
-            DataContext = ViewModel;
-        }
-
-        private async void FilterButton_Click(object sender, RoutedEventArgs e)
-        {
-            await App.ServiceProvider.GetService<AllTasks>().RefreshAsync();
+            ViewModel = this.SetDataContext<TasksPageViewModel>();
         }
     }
 }
